Twenty-four of the 26 `assets/` error fixtures the pinned Mojo compiles and runs are resolved. Four are respelled so the pin rejects the rule they pin (a run-time explicit-destroy condition, a moved `String` field, a genuinely ambiguous literal conversion, a write through an upgraded origin cast), and seven are recorded as deliberate strictness in `docs/non-goals.md` (hash-owned `for ref` writes, an out-of-place pointer offset, a dangling `Span` return, `TypeList` bounds, impure CTFE, the always-on stdlib assert), joined by a new non-trivial tracked `unsafe_take_pointee` fixture. The rest are implemented and verified against the pin on the VM and natively: reflective `Hashable`/`Writable` beside non-protocol overloads (with native reflective struct display), exact-source `@implicit` selection, raising `@implicit` constructors, `Codepoint(UInt8)`, the origin-cast upgrade that keeps an immutable capability, tracked trivial `unsafe_take_pointee`, union-origin `Pointer(to=ref)`, generic value aliases, comptime lambda calls and `String` values, fused-multiply-add float strided ranges, thin function-value array displays (fixing an element-call reference that let its receiver die first), and `x^` into a read parameter or receiver, which now lends instead of moving. Runtime `DType` and variadic-struct inference, plus five residues found on the way, are new roadmap checkboxes.

+++ assets/type_error/ambiguous_implicit_conversion.mojo
@@ -1,13 +1,15 @@
 # expect: ambiguous implicit conversion
+# A literal both converting constructors accept, and neither takes at its
+# default `Int`: no candidate is exact, so the conversion is ambiguous.
 struct Number:
     var value: Int
 
     @implicit
-    def __init__(out self, value: Int):
-        self.value = value
+    def __init__(out self, value: Int8):
+        self.value = Int(value)
 
     @implicit
-    def __init__(out self, value: UInt):
+    def __init__(out self, value: Int16):
         self.value = Int(value)
 
 def consume(value: Number):
